pyproject.toml
[tool.poetry]
name = "melody"
version = "0.1.0"
description = "An advanced rpm-ostree compose system"
authors = ["Lleyton Gray <lleyton@fyralabs.com>"]
license = "GPL3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
rich = "^12.5.1"
PyYAML = "^6.0"
anytree = "^2.8.0"

[tool.poetry.group.dev.dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
